+++ Blasphemous.LostDreams/Beads/RB501.cs
@@ -1,15 +1,36 @@
-﻿using Gameplay.GameControllers.AnimationBehaviours.Player.Attack;
+﻿using Blasphemous.Framework.Items;
+using Gameplay.GameControllers.AnimationBehaviours.Player.Attack;
 using Gameplay.GameControllers.Penitent.Abilities;
 using HarmonyLib;
 using UnityEngine;
 
-namespace Blasphemous.LostDreams.Effects;
+namespace Blasphemous.LostDreams.Beads;
+
+internal class RB501 : ModItemEffectOnEquip
+{
+    protected override void ApplyEffect()
+    {
+        DecreaseChangeTime = true;
+    }
+
+    protected override void RemoveEffect()
+    {
+        DecreaseChangeTime = false;
+    }
+
+    protected override void Update()
+    {
+        Main.LostDreams.LogWarning("Decrease: " + DecreaseChangeTime);
+    }
+
+    public static bool DecreaseChangeTime { get; private set; }
+}
 
 /// <summary>
 /// Reduce how much time to activate charge
 /// </summary>
 [HarmonyPatch(typeof(ChargedAttack), "SetChargingTimeByTier")]
-class ChargedAttack_SetTimer_Patch
+class RB501_ChargedAttack_SetTimer_Patch
 {
     public static void Postfix(ref float ____currentChargingTime)
     {
@@ -26,7 +47,7 @@ class ChargedAttack_SetTimer_Patch
 /// Increase speed of charging animation
 /// </summary>
 [HarmonyPatch(typeof(ChargedAttackBehaviour), "OnStateEnter")]
-class ChargedAnimation_Enter_Patch
+class RB501_ChargedAnimation_Enter_Patch
 {
     public static void Postfix(Animator animator)
     {
@@ -37,7 +58,7 @@ class ChargedAnimation_Enter_Patch
     }
 }
 [HarmonyPatch(typeof(ChargedAttackBehaviour), "OnStateExit")]
-class ChargedAnimation_Exit_Patch
+class RB501_ChargedAnimation_Exit_Patch
 {
     public static void Postfix(Animator animator)
     {
